Refactor to use ActiveRecord more
For each quiz question, I re-wrote it to use ActiveRecord to do the analysis, rather than using Ruby.

Along the way I found a pretty cool feature of ActiveRecord: counter cache. https://guides.rubyonrails.org/association_basics.html#options-for-belongs-to-counter-cache

Specifically I added a counter to Creator so every time a book with that creator is added, it adds 1 to the counter, and I added a counter to Publisher so every time a book is added with a given creator and publisher, it adds 1 to that counter.  This makes the code in command_line_interface really clean